+++ src/controllers/UserController.js
@@ -1,11 +1,7 @@
 import models from '../models';
 import { Authentication, SendEmail, HelperMethods } from '../utils';
 
-const {
-  Sequelize,
-  User,
-  LikedAccommodation
-} = models;
+const { User, } = models;
 
 /**
  * Class representing the user controller
@@ -372,54 +368,6 @@ class UserController {
       );
     } catch (error) {
       if (error.errors) return HelperMethods.sequelizeValidationError(res, error);
-      return HelperMethods.serverError(res);
-    }
-  }
-
-  /**
-   * Update user to like/unlike accommodation
-   * Route: POST: api/v1/
-   * @param {object} req - HTTP Request object
-   * @param {object} res - HTTP Response object
-   * @return {res} res - HTTP Response object
-   * @memberof UserController
-   */
-  static async likeAccommodation(req, res) {
-    try {
-      const { id: userId } = req.decoded;
-      const { like, accommodationId } = req.body;
-      const where = { userId, accommodationId };
-      const isAlreadyLiked = await LikedAccommodation.findOne({ where });
-      const success = isLiked => HelperMethods.requestSuccessful(
-        res,
-        {
-          success: true,
-          message: 'Update successful',
-          like: isLiked
-        },
-        200
-      );
-
-      if (isAlreadyLiked && like) return success(true);
-      if (!isAlreadyLiked && !like) return success(false);
-      if (like) {
-        await LikedAccommodation.create({ ...where });
-        return success(true);
-      }
-
-      if (!like) {
-        await LikedAccommodation.destroy({ where });
-        return success(false);
-      }
-    } catch (error) {
-      if (error instanceof Sequelize.ForeignKeyConstraintError) {
-        return HelperMethods.clientError(res, 'Accommodation does not exist', 404);
-      }
-
-      if (error instanceof Sequelize.DatabaseError) {
-        return HelperMethods.clientError(res, '"accommodationId" field is invalid', 400);
-      }
-
       return HelperMethods.serverError(res);
     }
   }
