+++ src/controllers/index.js
@@ -1,9 +1,13 @@
-/* eslint-disable import/no-cycle */
 import UserController from './UserController';
 import NotificationController from './NotificationController';
 import RequestController from './RequestController';
 import SocialController from './socialController';
+import AccommodationController from './AccommodationController';
 
 export {
-  UserController, RequestController, SocialController, NotificationController,
+  UserController,
+  RequestController,
+  SocialController,
+  NotificationController,
+  AccommodationController
 };
