+++ src/test/integrationTests/index.js
@@ -1,3 +1,4 @@
 import './requestController.test';
 import './userController.test';
 import './socialController.test';
+import './accommodationController.test';
